### 2019-03-11 #1314 @whedon
[PRE REVIEW]: VBLinLogit: Variational Bayesian linear and logistic regression
**Submitting author:** @jdrugo (<a href="http://orcid.org/0000-0002-7846-0408">Jan Drugowitsch</a>)
**Repository:** <a href="https://github.com/DrugowitschLab/VBLinLogit" target ="_blank">https://github.com/DrugowitschLab/VBLinLogit</a>
**Version:** v0.2
