+++ paper.bib
@@ -109,7 +109,7 @@
   pages = "211--244"
 }
 
-@article{Drugowitsch:2013
+@article{Drugowitsch:2013,
   author = "Jan Drugowitsch",
   title = "Variational Bayesian inference for linear and logistic regression",
   year = "2013",
