### 2019-03-11 #1314 @kyleniemeyer
@jdrugo looks like there is an error in your bibliography preventing compilation of the paper.
