### 2019-03-11 #1314 @kyleniemeyer
@eramirem and @arokem as the resident machine learning editors, could one of you handle this submission?
