### 2019-03-11 #1314 @jdrugo
@kyleniemeyer I have fixed the paper.bib typo upstream. How do I get whedon to pick up that change?
