### 2019-03-11 dec3694
Fix references in paper.md and paper.bib
