+++ paper.bib
@@ -83,7 +83,7 @@
   number = "1"
 }
 
-@article{Mackay:1992,
+@article{MacKay:1992,
   author = "David J. C. MacKay",
   title = "Bayesian Interpolation",
   journal = "Neural Computation",
