### 2019-03-11 #1314 @jdrugo
The article proof looks good now.
