### 2019-03-16 #1314 @labarba
Hi @jdrugo — we're looking to assign a handling editor for your submission (I am the Associated Editor in Chief on rotation this week). Before we move forward, I do have some comments on your paper.

From our [Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), the paper should contain:

- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
- A clear statement of need that illustrates the purpose of the software
- Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it

I want to emphasize the need for explaining the software to a diverse audience: currently, I don't see this in the paper; the second paragraph is especially jargon-heavy. You also don't mention if this software has been used in any research projects.

Note also the section on "novelty":

> Submissions that implement solutions already solved in other software packages are accepted into JOSS provided that they meet the criteria listed above and cite prior similar work.

Can you comment about existing related functionality in Matlab? For example, the Matlab documentation has a section on [Bayesian regression models](https://www.mathworks.com/help/econ/bayesian-linear-regression-models.html). How does your software fill a gap in this ecosystem?
