### 2019-03-16 #1314 @labarba
@jdrugo — Can you check your bibliography entries and add a DOI whenever one exists? Our helpful bot `whedon` has given some clues above for possible missing DOIs (it's not perfect, so some could be flukes).
