### 2019-03-20 bb02c70
Add DOI or ISBN to bibliography whenever available
