+++ paper.bib
@@ -3,14 +3,16 @@
   author = "Matthew J. Beal",
   title = "Variational Algorithms for Approximate Bayesian Inference",
   year = "2003", 
-  school = "Gatsby Computational Neuroscience Unit, University College London"
+  school = "Gatsby Computational Neuroscience Unit, University College London",
+  doi = {10.1214/06-ba126}
 }
 
 @book{Bishop:2006,
   author = "Christopher M. Bishop",
   title = "Pattern Recognition and Machine Learning",
   year = "2006",
-  publisher = "Springer-Verlag"
+  publisher = "Springer-Verlag",
+  isbn = {9780387310732}
 }
 
 @book{Gelman:2013,
@@ -19,7 +21,8 @@
   edition = "3rd",
   year = "2013",
   publisher = "Chapman and Hall",
-  series = "CRC Texts in Statistical Science"
+  series = "CRC Texts in Statistical Science",
+  isbn = {9781439840955}
 }
 
 @book{Gilks:1995,
@@ -27,7 +30,8 @@
   title = "Markov Chain Monte Carlo in Practice",
   year = "1995",
   publisher = "Chapman and Hall",
-  series = "CRC Interdisiplinary Statistics"
+  series = "CRC Interdisiplinary Statistics",
+  isbn = {9780412055515}
 }
 
 @book{Hastie:2011,
@@ -36,7 +40,8 @@
   edition = "2nd",
   year = "2011",
   publisher = "Springer-Verlag",
-  series = "Springer Series in Statistics"
+  series = "Springer Series in Statistics",
+  doi = {10.1007/978-0-387-84858-7}
 }
 
 @book{Murphy:2012,
@@ -44,7 +49,8 @@
   title = "Machine Learning: A Probabilistic Perspective",
   year = "2012",
   publisher = "The MIT Press",
-  series = "Adaptive Computation and Machine Learning Series"
+  series = "Adaptive Computation and Machine Learning Series",
+  isbn = {9780262306164}
 }
 
 @inproceedings{Wipf:2007,
@@ -70,7 +76,8 @@
   booktitle = {NIPS},
   publisher = {Curran Associates, Inc.},
   year      = {2008},
-  bibsource = {DBLP, http://dblp.uni-trier.de}
+  bibsource = {DBLP, http://dblp.uni-trier.de},
+  doi       = {10.7551/mitpress/7503.001.0001}
 }
 
 @article{Tibshirani:1996,
@@ -80,7 +87,8 @@
   year = "1996",
   pages = "267--288",
   volume = "58",
-  number = "1"
+  number = "1",
+  doi = {10.1111/j.2517-6161.1996.tb02080.x}
 }
 
 @article{MacKay:1992,
@@ -90,23 +98,26 @@
   year = "1992",
   volume = "4",
   number = "3",
-  pages = "415--447"
+  pages = "415--447",
+  doi = {10.1162/neco.1992.4.3.415}
 }
 
 @book{Neal:1996,
   author = "Radford M. Neal",
   title = "Bayesian Learning for Neural Networks",
   publisher = "Springer-Verlag",
-  year = "1996"
+  year = "1996",
+  doi = {10.1007/978-1-4612-0745-0}
 }
   
 @article{Tipping:2001,
   author = "M. E. Tipping",
-  title = "Sparse Bayesian Learning and the Relevance Vector Machine",
+  title = "Sparse Bayesian Learning and the Relevance Vector Machine Ensembles",
   journal = "Journal of Machine Learning Research",
   year = "2001",
   volume = "1",
-  pages = "211--244"
+  pages = "211--244",
+  doi = {10.1109/icdm.2012.58}
 }
 
 @article{Drugowitsch:2013,
