### 2019-03-20 715c980
Simply paper.md and add additional details on novelty
