+++ paper.md
@@ -18,10 +18,20 @@ bibliography: paper.bib
 
 # Summary
 
-Linear and logistic regression are essential workhorses of statistical analysis, whose Bayesian treatment has received much recent attention [@Gelman:2013; @Bishop:2006; @Murphy:2012; @Hastie:2011]. These allow specifying the a-priori uncertainty and infer a-posteriori uncertainty about regression coefficients explicitly and hierarchically, by, for example, specifying how uncertain we are a-priori that these coefficients are small. However, Bayesian inference in such hierarchical models quickly becomes intractable, such that recent effort has focused on approximate inference, like Markov Chain Monte Carlo methods [@Gilks:1995], or variational Bayesian approximation [@Beal:2003; @Bishop:2006; @Murphy:2012).
+Linear and logistic regression are essential workhorses of statistical analysis, whose Bayesian treatment has received much recent attention [@Gelman:2013; @Bishop:2006; @Murphy:2012; @Hastie:2011]. Using Bayesian statistics for linear and logistic regression allows specifying prior beliefs over certain model parameters, which makes it particularly useful for small and/or high-dimensional datasets. Bayesian regression furthermore provides an estimate of its uncertainty in the estimated regression coefficient, as well as uncertainty when making predictions arising from the regression. Both are again particularly important for small and/or high-dimensional datasets, as these might result in highly uncertain predictions, and allow the user to be explicit about this uncertainty.
 
-``VBLinLogit`` is a MATLAB library that provides an variational Bayesian implementation of Bayesian hierarchical models for both linear and logistic regression. Both models include a variant with automatic relevance determination (ARD), which consists of assigning an individual hyper-prior to each regression coefficient separately. These  hyper-priors are adjusted to eventually prune irrelevant coefficients [@Wipf:2007] without the need for a separate validation set, unlike comparable sparsity-inducing methods like the Lasso [@Tibshirani:1996]. Bishop [-@Bishop:2006] describes ARD only in the  context of type-II maximum likelihood [@MacKay:1992; @Neal:1996; @Tipping:2001], where it (hyper-)parameters are tuned by maximizing the marginal likelihood (or model evidence). Here, instead, we provide implementations for the full Bayesian treatment, that finds the ARD hyper-posteriors by variational Bayesian inference.
+``VBLinLogit`` is a MATLAB library that provides an variational Bayesian implementation of Bayesian models for both linear and logistic regression. It uses variational Bayesian inference [@Beal:2003; @Bishop:2006; @Murphy:2012] --- a method for approximating Bayesian computations --- as the Bayesian computations underlying the used models would otherwise be intractable. It is significantly faster than Markov Chain Monte Carlo (MCMC) methods [@Gilks:1995] --- another form of approximate Bayesian inference --- which makes it applicable to high-dimensional problems for which standard MCMC might be too slow.
 
-The scripts encompassing the library were written to be light-weight and thus do not depend on external libraries. Details about the variational Bayes derivation and use of the library can be found in Drugowitsch [-@Drugowitsch:2013].
+A specific regression variant implemented by this library is automatic relevance determination (ARD), which uses a model that automatically determines which data dimensions are relevant for the regression, discarding the others [@Wipf:2007]. It does so without a separate "validation set", as would be required by alternative methods, like the Lasso [@Tibshirani:1996]. Therefore, it can be used when the small size of the dataset makes creating such a separate "validation set" prohibitive.
+
[…]
