+++ paper.bib
@@ -102,6 +102,17 @@
   doi = {10.1162/neco.1992.4.3.415}
 }
 
+@article{George:1993,
+    author = "Edward I. George and Robert E. {McCulloch}",
+    title = "Variable Selection via Gibbs Sampling",
+    journal = "Journal of the American Statistical Association",
+    year = "1993",
+    volume = "88",
+    number = "423",
+    pages = "881-889",
+    doi = {10.1080/01621459.1993.10476353}
+}
+
 @book{Neal:1996,
   author = "Radford M. Neal",
   title = "Bayesian Learning for Neural Networks",
@@ -127,4 +138,44 @@
   Archiveprefix = {arXiv},
   Eprint = {1310.5438},
   Journal = {ArXiv e-prints},
+}
+
+@article{Kanitscheider:2015,
+    author = "Ingmar Kanitscheider and Ruben Coen-Cagli and Adam Kohn and Alexandre Pouget",
+    title = "Measuring Fisher Information Accurately in Correlated Neural Populations",
+    journal = "{PLoS} Computational Biology",
+    volume = "11",
+    number = "6",
+    pages = "e1004218",
+    doi = {10.1371/journal.pcbi.1004218}
+}
+
+@article{Oh:2016,
+  author = "Hanna Oh and Jeffrey M. Beck and Pingping Zhu and Marc A. Sommer and Silvia Ferrari and Tobias Egner",
+  title = "Satisficing in Split-Second Decision Making is Characterized by Strategic Cue Discounting",
+  journal = "Journal of Experimental Psychology: Learning, Memory, and Cognition",
+  year = "2016",
+  volume = "42",
+  number = "12",
+  pages = "1937--1956",
+  doi = {10.1037/xlm0000284}
+}
+
+@article{Wang:2017,
+  author = "Shaoming Wang and Bob Rehder",
+  title = "Multi-attribute Decision-making is Best Characterized by an Attribute-Wide Reinforcement Learning Model",
+  journal = "bioRxiv",
+  year = "2017",
+  dio = {10.1101/234732}
+}
+
+@article{Ruxanda:2018,
+  author = "Gheorghe Ruxanda and Catalina Zahfir and Andreea Muraru",
+  title = "Predicting financial distress from Romanian companies",
+  journal = "Technological and Economic Development of Economy",
+  year = "2018",
+  volume = "24",
+  number = "6",
+  pages = "2318--2337",
+  doi = {10.3846/tede.2018.6736}
 }
