### 2019-03-20 5f91864
Fix typos, spelling, grammar, in paper.md and paper.bib
