+++ paper.md
@@ -18,20 +18,20 @@ bibliography: paper.bib
 
 # Summary
 
-Linear and logistic regression are essential workhorses of statistical analysis, whose Bayesian treatment has received much recent attention [@Gelman:2013; @Bishop:2006; @Murphy:2012; @Hastie:2011]. Using Bayesian statistics for linear and logistic regression allows specifying prior beliefs over certain model parameters, which makes it particularly useful for small and/or high-dimensional datasets. Bayesian regression furthermore provides an estimate of its uncertainty in the estimated regression coefficient, as well as uncertainty when making predictions arising from the regression. Both are again particularly important for small and/or high-dimensional datasets, as these might result in highly uncertain predictions, and allow the user to be explicit about this uncertainty.
+Linear and logistic regression are essential workhorses of statistical analysis, whose Bayesian treatment has received much recent attention [@Gelman:2013; @Bishop:2006; @Murphy:2012; @Hastie:2011]. Using Bayesian statistics for linear and logistic regression allows specifying prior beliefs over certain model parameters, which makes it particularly useful for small and/or high-dimensional datasets. Bayesian regression furthermore provides an estimate of the uncertainty about estimated regression coefficient, as well as uncertainty about predictions arising from the regression. Both are again particularly important for small and/or high-dimensional datasets, as the use of such data might result in highly uncertain predictions, and allow the user to be explicit about this uncertainty.
 
-``VBLinLogit`` is a MATLAB library that provides an variational Bayesian implementation of Bayesian models for both linear and logistic regression. It uses variational Bayesian inference [@Beal:2003; @Bishop:2006; @Murphy:2012] --- a method for approximating Bayesian computations --- as the Bayesian computations underlying the used models would otherwise be intractable. It is significantly faster than Markov Chain Monte Carlo (MCMC) methods [@Gilks:1995] --- another form of approximate Bayesian inference --- which makes it applicable to high-dimensional problems for which standard MCMC might be too slow.
+``VBLinLogit`` is a MATLAB library that provides a variational Bayesian implementation of Bayesian models for both linear and logistic regression. It uses variational Bayesian inference [@Beal:2003; @Bishop:2006; @Murphy:2012] as a method for approximating Bayesian computations, as these computations would otherwise be intractable for the used regression models. It is significantly faster than Markov Chain Monte Carlo (MCMC) methods [@Gilks:1995], another form of approximate Bayesian inference, which makes it applicable to high-dimensional problems for which standard MCMC might be too slow.
 
-A specific regression variant implemented by this library is automatic relevance determination (ARD), which uses a model that automatically determines which data dimensions are relevant for the regression, discarding the others [@Wipf:2007]. It does so without a separate "validation set", as would be required by alternative methods, like the Lasso [@Tibshirani:1996]. Therefore, it can be used when the small size of the dataset makes creating such a separate "validation set" prohibitive.
+A specific regression variant implemented by this library is automatic relevance determination (ARD), which uses a model that automatically determines which data dimensions are relevant for the regression, discarding the others [@Wipf:2007]. It does so without a separate "validation set", as would be required by alternative methods, like the Lasso [@Tibshirani:1996]. Therefore, it can be used when the small size of the dataset makes the use of such a separate "validation set" prohibitive.
 
[…]
