+++ paper.bib
@@ -144,6 +144,7 @@
     author = "Ingmar Kanitscheider and Ruben Coen-Cagli and Adam Kohn and Alexandre Pouget",
     title = "Measuring Fisher Information Accurately in Correlated Neural Populations",
     journal = "{PLoS} Computational Biology",
+    year = "2015",
     volume = "11",
     number = "6",
     pages = "e1004218",
@@ -166,7 +167,7 @@
   title = "Multi-attribute Decision-making is Best Characterized by an Attribute-Wide Reinforcement Learning Model",
   journal = "bioRxiv",
   year = "2017",
-  dio = {10.1101/234732}
+  doi = {10.1101/234732}
 }
 
 @article{Ruxanda:2018,
