### 2019-03-20 #1314 @jdrugo
@labarba -- Thank you for your comments. I have now updated the paper accordingly, and have added DOIs whenever available.
