### 2019-03-20 #1314 @labarba
Did you also address my other comment, above?
https://github.com/openjournals/joss-reviews/issues/1314#issuecomment-473580071
