### 2019-03-20 #1314 @jdrugo
@labarba - yes, the updated version is an attempt to address these comments. I didn't go so far as to explain what linear and logistic regression are (I hope this knowledge can be assumed), but tried to simplify the summary.
