### 2019-03-30 #1314 @usethedata
@jdrugo My apologies for my delays in getting back to this submission.  My work life got a bit intense.  The revised manuscript looks good from my perspective.  Let me see about getting some reviewers assigned to this.
