### 2019-05-02 DrugowitschLab/VBLinLogit#1 @ManuelaS
JOSS review: Installation
Add installation steps in the ReadMe
