### 2019-05-02 DrugowitschLab/VBLinLogit#2 @ManuelaS
JOSS review: Requirements
Detail requirements in the ReadMe
