### 2019-05-02 DrugowitschLab/VBLinLogit#3 @ManuelaS
JOSS review: Documentation API
Improve documentation API
