### 2019-05-02 DrugowitschLab/VBLinLogit#4 @ManuelaS
JOSS review: examples
Flesh out example descriptions and also include in the ReadMe
