### 2019-05-02 DrugowitschLab/VBLinLogit#5 @ManuelaS
JOSS review: Automated tests
Add automated unit tests for each regression implementation
