### 2019-05-02 DrugowitschLab/VBLinLogit#6 @ManuelaS
JOSS review: Contributing guidelines
Improve contributing guidelines
