### 2019-05-02 #1359 @ManuelaS
Hi @jdrugo,

I have started to review `VBLinLogit`. I really enjoyed reviewing this library, I like its simplicity and I will definitely try it out on some of my work in the future!
Based on JOSS reviewing criteria, I have the following comments so far:

### Software:

- [ ]  **Installation:** Installation steps should be documented in the ReadMe as some potential users may have never used 3rd party code in MATLAB before (issue [here](https://github.com/DrugowitschLab/VBLinLogit/issues/1)).

- [ ]  **Requirements:** (issue [here](https://github.com/DrugowitschLab/VBLinLogit/issues/2))
   - Detail MATLAB version the library was developed/tested on and expect to work on. I am using MATLAB 2014b and worked smoothly, but I believe this library would work also in versions a lot older?
   - Some examples require the Statistics toolbox to run top to bottom and it may be worthwhile highlighting what the users without the toolbox would need to do to work around it or what they would miss out on (the `regress` function is in the Statistics toolbox).
   - Would this package work as expected in Octave? If it did, the users base would be greatly expanded and may be worthwhile mentioning explicitly in the ReadMe.

- [ ]  **Documentation API:** (issue [here](https://github.com/DrugowitschLab/VBLinLogit/issues/3))
I really appreciate that each function includes the math behind the implementation in the doc string! Though I would strive to also document inputs and outputs of the function from a "software perspective" (data types, etc.) so that users can quickly look up *how* to use the functions (without necessarily need to understand the theory behind it).

- [ ]  **Example usage:** (issue [here](https://github.com/DrugowitschLab/VBLinLogit/issues/4))
Examples are included and bundled in `vb_examples.m` (though only 6 out of the 8 examples included in the repository are called?). Examples ran as expected and produced the figures as shown in [arxiv:1310.5438](https://arxiv.org/abs/1310.5438). I would suggest:
   - Adding instructions on how to run the set of examples explicitly in the ReadMe;
   - Selecting a couple of examples and including them directly in the ReadMe. I think it would help the reader assess what the library can do and whether or not it would suit their needs.  Additionally, I would strive to expand the descriptions for the examples that is a bit "thin" at the moment bearing in mind a non-specialised audience.

- [ ]  **Automated tests:** I did not find any unit tests. I would suggest  (issue [here](https://github.com/DrugowitschLab/VBLinLogit/issues/5)):
    - Adding automated tests that cover each regression implementations and bundle them in a unit test suite;
    - Adding instructions on how to run the tests in the ReadMe, so that users can verify that the installation was successful and can use those tests as base if they decide to implement additional methods.

- [ ] **Contributing guidelines**: (issue [here](https://github.com/DrugowitschLab/VBLinLogit/issues/6))
While the ReadMe remains the readers to get in touch via GitHub for bug reports and contributions, I would suggest:
   - Adding explicitly a CONTRIBUTING.md file (see for example https://www.contributor-covenant.org/) to encourage users to participate with PR requests to further develop the library;
   - Adding instructions on how to get in touch to seek support (and not necessarily report bugs/request features as per point 3 in the "Contributing guidelines" section).

### Paper
- I really enjoyed reading this manuscript and its extended version in [arxiv](https://arxiv.org/abs/1310.5438). Though I wonder if the second part of the paper could be made more accessible to distill the advantages/disadvantages of this particular implementation geared towards a non-specialized audience and maybe just
remind the more mathematically-oriented readers to the [arxiv paper](https://arxiv.org/abs/1310.5438)?
Feel free to ignore this comment, it may be beyond scope of the JOSS submission (@usethedata: what is your take on this?) but I believe it would be beneficial to provide the reader with an *easy-to-read* account of how VBLinLogit fits in the Bayesian linear/logistic
regression ecosystem (including non-MATLAB based implementations). I believe a simple description would go a long way to clarify to the reader whether this library would fit their use case/research. I would also suggest including this information quite prominently on the ReadMe.
- There are a couple of DOIs missing (see above output from `@whedon check references`)
