### 2019-06-10 08e46a8
Update code to run in Octave and without the stats toolbox
All calls to regress(.) in the linear regression scripts were made
optional. In case regress(.) isn't available, no confidence intervals
are plotted.

To additionally ensure Octave compatibility, calls to rng(.) are
replaced by calls to rand("state", .).
