### 2019-06-10 DrugowitschLab/VBLinLogit#2 @jdrugo
Fixed in 08e46a8.
