### 2019-06-10 204a5c4
Split files into src and example folders
As a preparation to add tests, the files are now split into src and
examples files, and the documentation has been updated accordingly.
