+++ examples/vb_examples.m
@@ -1,9 +1,11 @@
 %% script tu run all vb_{linar,logit}_example_* scripts
 
 example_scripts = {...
+    'vb_linear_example.m', ...
     'vb_linear_example_highdim.m', ...
     'vb_linear_example_sparse.m', ...
     'vb_linear_example_modelsel.m', ...
+    'vb_logit_example.m', ...
     'vb_logit_example_coeff.m', ...
     'vb_logit_example_highdim.m', ...
     'vb_logit_example_modelsel.m'};
