+++ examples/vb_linear_example.m
@@ -1,18 +1,25 @@
-% script demonstrating the use of vb_linear_fit.m and
+%% simple script demonstrating the use of vb_linear_fit.m and
 % vb_linear_fit_ard.m
 %
 % Copyright (c) 2013, Jan Drugowitsch
 % All rights reserved.
 % See the file LICENSE for licensing information.
 
-% dimensionality, number of data points, noise
+
+%% set RNG seed
+isOctave = exist('OCTAVE_VERSION', 'builtin') ~= 0;
+if isOctave, rand("state", 4); else, rng(4); end
+
+
+%% dimensionality, number of data points, noise
 d = 4;
 d_extra = 10;
 N = 50;
 N_cv = 50;
 tau = 1;
 
-% random weight vector & predictions
+
+%% random weight vector & predictions
 w = randn(d, 1);
 x = rand(N, 1);
 X = bsxfun(@power, x, 0:(d-1));
@@ -25,13 +32,15 @@ x_pred = linspace(0, 1, 100)';
 X_pred = bsxfun(@power, x_pred, 0:(d-1));
 X_pred_ext = bsxfun(@power, x_pred, 0:(d+d_extra-1));
 
-% weights by maximum likelihood
-w_ML = regress(y, X);
-w_ML_ext = regress(y, X_ext);
+
+%% weights by maximum likelihood
+w_ML = X \ y;
+w_ML_ext = X_ext \ y;
 y_ML = X_ext * w_ML_ext;
 y_ML_cv = X_cv * w_ML_ext;
 
-% weights and predictions by variational bayes (only on extended space)
+
+%% weights and predictions by variational bayes (only on extended space)
 [w_vb V_vb invV_vb logdetV_vb an_vb bn_vb] = vb_linear_fit(X_ext, y);
 [y_vb lam_vb nu_vb] = vb_linear_pred(X_pred_ext, w_vb, V_vb, an_vb, bn_vb);
 % standard deviation of Student's t with parameters lam_vb and nu_vb
@@ -42,7 +51,8 @@ y_vb_sd = sqrt(nu_vb ./ (lam_vb .* (nu_vb - 2)));
 [y_vb_ard lam_vb nu_vb] = vb_linear_pred(X_pred_ext, w_vb_ard, V_vb, an_vb, bn_vb);
 y_vb_ard_sd = sqrt(nu_vb ./ (lam_vb .* (nu_vb - 2)));
 
-% plot fits
+
+%% plot fits
 figure; set(0,'defaultlinelinewidth', 1.5);  hold on;
 plot(x_pred, X_pred * w, 'k-');
 plot(x, y, 'k+');
