+++ examples/vb_logit_example.m
@@ -1,18 +1,24 @@
-% script demonstrating the use of bayes_logit_fit.m and
+%% simple script demonstrating the use of bayes_logit_fit.m and
 % bayes_logit_fit_ard.m
 %
 % Copyright (c) 2013, Jan Drugowitsch
 % All rights reserved.
 % See the file LICENSE for licensing information.
 
-% dimensionality, number of data points
+
+%% set RNG seed
+isOctave = exist('OCTAVE_VERSION', 'builtin') ~= 0;
+if isOctave, rand("state", 0); else, rng(0); end
+
+
+%% dimensionality, number of data points
 d = 3;
 d_extra = 10;
 N = 100;
 N_cv = 100;
 
 
-% random weight vector & predictions
+%% random weight vector & predictions
 w = randn(d, 1);
 X = [ones(N, 1) randn(N, d-1)];
 X_ext = [X randn(N, d_extra)];
@@ -23,7 +29,8 @@ y = 2 * (rand(N, 1) < p_y) - 1;
 y1 = (y == 1);
 y_cv = 2 * (rand(N_cv, 1) < 1 ./ (1 + exp(- X_cv(:, 1:d) * w))) - 1;
 
-% Fisher Linear Discriminant
+
+%% Fisher Linear Discriminant
 w_LD = NaN(d, 1);
 w_LD(2:end) = (cov(X(~y1, 2:end)) + cov(X(y1, 2:end))) \ ...
               (mean(X(y1, 2:end))' - mean(X(~y1, 2:end))');
@@ -36,7 +43,7 @@ y_LD = 2 * (X_ext * w_LD_ext > 0) - 1;
 y_LD_cv = 2 * (X_cv * w_LD_ext > 0) - 1;
 
 
-% weights and predictions by variational Bayes
+%% weights and predictions by variational Bayes
 [w_vb, V_vb, invV_vb, logdetV_vb, E_a_vb, L_vb] = vb_logit_fit(X_ext, y);
 p_y_vb = vb_logit_pred(X_ext, w_vb, V_vb, invV_vb);
 y_vb = 2 * (p_y_vb > 0.5) - 1;
@@ -50,7 +57,8 @@ y_vb_ard = 2 * (p_y_vb_ard > 0.5) - 1;
 p_y_vb_ard_cv = vb_logit_pred(X_cv, w_vb_ard, V_vb, invV_vb);
 y_vb_ard_cv = 2 * (p_y_vb_ard_cv > 0.5) - 1;
 
-% plot data and discriminating hyperplane
+
+%% plot data and discriminating hyperplane
 figure;  hold on;
 plot(X(~y1, 2), X(~y1, 3), 'b+');
 plot(X(y1, 2), X(y1, 3), 'r+');
