+++ examples/README.md
@@ -1,0 +1,7 @@
+VBLinLogit example scripts
+==========================
+
+The files starting `vb_linear` provide examples for the use of variational Bayesian linear regression. The files starting in `vb_logit` provide examples for variational Bayesian logistic regression. Calling `vb_examples` results in running all example scripts in this folder.
+
+The scripts `vb_linear_example_highdim`, `vb_linear_example_modelsel`, `vb_linear_example_sparse`, `vb_logit_example_coeff`, `vb_logit_example_highdim` and `vb_logit_example_modelsel` reproduce the figures in *Variational Bayesian
+inference for linear and logistic regression*, [arxiv:1310.5438](http://arxiv.org/abs/1310.5438) [stat.ML].
