### 2019-06-11 9992495
Add more documentation to examples, fixes #4
Expanded README.md to describe individual examples, made sure that all
examples run in Octave, and added more documentation/comments to
individual sample scripts.
