+++ examples/vb_linear_example_sparse.m
@@ -1,5 +1,14 @@
 %% sparse linear regression example for vb_linear_*
 %
+% This example demonstrates the ability of automated relevance
+% determination (ARD) to detect and ignore irrelevant input dimensions. The
+% script shows this by generating an input -> output mapping with a
+% D-dimensional input space, of which only a small subset of D_eff
+% dimensions determine the output. It then compares variational Bayesian
+% linear regression without and with ARD, and a least-square estimate, and
+% shows that the variant with ARD is better able to estimate the regression
+% coefficients, and also provides lower-error predictions on the test set. 
+%
 % Copyright (c) 2014, Jan Drugowitsch
 % All rights reserved.
 % See the file LICENSE for licensing information.
@@ -7,16 +16,17 @@
 
 %% set RNG seed and plot limits to re-produce arXiv figures (in MATLAB)
 isOctave = exist('OCTAVE_VERSION', 'builtin') ~= 0;
-if isOctave, rand("state", 0); else, rng(0); end
+if isOctave, rand("state", 0); randn("state", 0); else, rng(0); end
 wlims = [-5 5];
 ylims = [-15 15];
 
 
 %% settings
-D = 1000;
-D_eff = 100;
-N = 500;
-N_test = 50;
+D = 1000;        % full input dimensionality
+D_eff = 100;     % number of effective input dimensions
+N = 500;         % number of training set examples
+N_test = 50;     % number of test set examples
+
 % create data
 w = [randn(D_eff, 1); zeros(D - D_eff, 1)];
 X = rand(N, D) - 0.5;
@@ -26,15 +36,15 @@ y_test = X_test * w + randn(N_test, 1);
 
 
 %% preform regression and make predictions
-% variational bayes linear regression
+% variational bayes linear regression, train & test-set predictions
 [w_VB, V_VB, ~, ~, an_VB, bn_VB] = vb_linear_fit(X, y);
 y_VB = vb_linear_pred(X, w_VB, V_VB, an_VB, bn_VB);
 [y_test_VB, lam_VB, nu_VB] = vb_linear_pred(X_test, w_VB, V_VB, an_VB, bn_VB);
-% variational bayes linear regression with ARD
+% variational bayes linear regression with ARD, train & test-set predictions
 [w_VB2, V_VB2, ~, ~, an_VB2, bn_VB2] = vb_linear_fit_ard(X, y);
 y_VB2 = vb_linear_pred(X, w_VB2, V_VB2, an_VB2, bn_VB2);
 [y_test_VB2, lam_VB2, nu_VB2] = vb_linear_pred(X_test, w_VB2, V_VB2, an_VB2, bn_VB2);
-% maximum likelihood
+% maximum likelihood, train & test-set predictions
 if exist('regress','file') == 2
     [w_ML, wint_ML] = regress(y, X);
 else
@@ -42,7 +52,7 @@ else
 end
 y_ML = X * w_ML;
 y_test_ML = X_test * w_ML;
-% train and test set error
+% output train and test set error
 fprintf('Training set MSE: ML = %f, VB = %f, VB w/ ARD = %f\n', ...
         mean((y - y_ML).^2), mean((y - y_VB).^2), mean((y - y_VB2).^2));
[…]
