+++ examples/vb_logit_example_highdim.m
@@ -1,5 +1,15 @@
 %% high-dimensional logitstic regression, using vb_logit_*
 %
+% This example demonstrates the ability of automated relevance
+% determination (ARD) to detect and ignore irrelevant input dimensions for
+% logistic regression. The script shows this by generating an input ->
+% output mapping with a D-dimensional input space, of which only a small
+% subset of D_eff dimensions determine the output class. It then compares
+% variational Bayesian linear regression without and with ARD, and linear
+% discriminant analysis, and shows that the variant with ARD is better able
+% to estimate the regression coefficients, and also provides lower-error
+% predictions on the test set.
+%
 % Copyright (c) 2014, Jan Drugowitsch
 % All rights reserved.
 % See the file LICENSE for licensing information.
@@ -7,27 +17,30 @@
 
 %% set RNG seed and plot limits to re-produce arXiv figures (in MATLAB)
 isOctave = exist('OCTAVE_VERSION', 'builtin') ~= 0;
-if isOctave, rand("state", 0); else, rng(0); end
+if isOctave, rand("state", 0); randn("state", 0); else, rng(0); end
 wlims = [-4 4];
 
 
 %% settings
-D = 1000;
-D_eff = 100;
-N = 2000;
-N_test = 10000;
-N_plot = 200;
+D = 1000;        % full input dimensionality
+D_eff = 100;     % number of effective input dimensions
+N = 2000;        % number of training set examples
+N_test = 10000;  % number of test set examples
+N_plot = 200;    % number of examples to plot
+
 % generate data
 w = [randn(D_eff, 1); zeros(D - D_eff, 1)];
+% inputs for train & test set
 X = rand(N, D) - 0.5;
 X_test = rand(N_test, D) - 0.5;
+% output probabilities and samples for train & test set
 py = 1 ./ (1 + exp(- X * w));
 y = 2 * (rand(N, 1) < py) - 1;
 py_test = 1 ./ (1 + exp(-X_test * w));
 y_test = 2 * (rand(N_test, 1) < py_test) - 1;
 
 
-%% estimate coefficients, form predictions
+%% estimate coefficients, form predictions for train & test set
 % VB, no ARD
 fprintf('Variational Bayesian estimation, no ARD\n');
 [w_VB, V_VB, invV_VB] = vb_logit_fit(X, y);
@@ -43,20 +56,20 @@ fprintf('Variational Bayesian estimation, with ARD\n');
 [w_VB2, V_VB2, invV_VB2] = vb_logit_fit_ard(X, y);
 py_VB2 = vb_logit_pred(X, w_VB2, V_VB2, invV_VB2);
 py_test_VB2 = vb_logit_pred(X_test, w_VB2, V_VB2, invV_VB2);
-% Fisher linear discriminant
-fprintf('Fisher linear discriminant\n');
+% Fisher linear discriminant analysis
+fprintf('Fisher linear discriminant analysis\n');
 y1 = y == 1;
 w_LD = (cov(X(y1, :)) + cov(X(~y1, :))) \ ...
        (mean(X(y1, :))' - mean(X(~y1, :))');
[…]

[…]
