### 2019-06-11 dfe19ad
Refine contributing guidelines, fixes #6
