### 2019-06-11 f5d30d5
Restructure per-function documentation, add API to README.md, fixes #3
