### 2019-06-14 6e3da8b
Add unit tests, fixes #5
Added unit tests for each function in the src folder, and one script
that calls all of them.
