+++ test/vb_tests.m
@@ -1,0 +1,18 @@
+%% script to run all vb_{linar,logit}_*_test scripts
+
+test_scripts = {...
+    'vb_linear_fit_test.m',...
+    'vb_linear_fit_ard_test.m',...
+    'vb_linear_pred_test.m',...
+    'vb_logit_fit_test.m',...
+    'vb_logit_fit_ard_test.m',...
+    'vb_logit_fit_iter_test.m',...
+    'vb_logit_pred_test.m',...
+    'vb_logit_pred_incr_test.m'};
+
+for i = 1:length(test_scripts)
+    script_name = test_scripts{i};
+    fprintf('Running %s\n', script_name);
+    run(script_name);
+    fprintf('\n');
+end
