### 2019-06-15 bc7b891
Fix README.md and update JOSS paper
Added a simpler explanation of regularization to the JOSS paper,
and fixed one DOI.
