+++ paper.bib
@@ -31,7 +31,8 @@
   year = "1995",
   publisher = "Chapman and Hall",
   series = "CRC Interdisiplinary Statistics",
-  isbn = {9780412055515}
+  isbn = {9780412055515},
+  doi = {10.1080/09332480.2014.914768}
 }
 
 @book{Hastie:2011,
