### 2019-06-15 #1359 @jdrugo
@ManuelaS & @labarba, I hope to have addressed all the points in the review. The biggest changes are

* restructuring the code into different folder. The addition of tests would have made the number of files in the root folder too large.
* significantly improving README.md, adding more detailed installation instructions, requirements, API documentation.
* improving the documentation in the headers of the various scripts in `src`
* adding separate CONTRIBUTING.md.
* adding tests.
* Adding a couple of paragraphs to the paper to give a simpler explanation of regularization, and the difference between different regularization approaches.
* Added one DOI. I couldn't find a DOI for the second book; the one provided by whedon is wrong (it's a review of the book).

I don't know how you handle versions, but I plan on bumping the version to v0.3 once the review is done. I will need a new version number as the script line numbers have changed, and I need to update them in the arXiv paper.
