### 2019-06-16 #1359 @usethedata
@jdrugo Thanks.  It's not problem to update the version.  I've not done a lot of papers as editor, but it's happened with most of the ones I've handled.  On the DOI, that's fine.  Whedon's suggestions are not always correct.  I'll check out the updated code and go through my review.  We had some challenges finding reviewers, so I'm doing a review as well as being the editor on this one.
