### 2019-06-16 #1359 @usethedata
I confirm that the suggested DOI (https://doi.org/10.1080/09332480.2014.914768) for "Machine Learning: A Probabilistic Perspective" is incorrect.  I do suggest using the arXiv persistent identifier URL, however (https://arxiv.org/abs/1310.5438).
