### 2019-06-16 c159e0d
Add URL to arXiv reference, fixes #8
