+++ paper.bib
@@ -139,6 +139,7 @@
   Archiveprefix = {arXiv},
   Eprint = {1310.5438},
   Journal = {ArXiv e-prints},
+  url = {https://arxiv.org/abs/1310.5438}
 }
 
 @article{Kanitscheider:2015,
