### 2019-06-23 #1359 @ManuelaS
Hi @jdrugo  & @usethedata, I checked all the remaining boxes in the review and I am happy to recommend it to be accepted from my side.

Note that I checked the box regarding the version as well, but it will need to be bumped to include the last changes.
