### 2019-06-23 #1359 @usethedata
@jdrugo:  A few items.

1. Please check the fifth reference:  Gilks, W. R., Richardson, S., & Spiegelhalter, D. (Eds.). (1995). Markov chain monte carlo in
practice. CRC interdisiplinary statistics. Chapman; Hall. doi:10.1080/09332480.2014.914768

When I follow that DOI, I'm not getting the work you're referencing.

2. Please also check the last reference: Wipf, D. P., & Nagarajan, S. S. (2007). A new view of automatic relevance determination.
In J. C. Platt, D. Koller, Y. Singer, & S. T. Roweis (Eds.), NIPS. Curran Associates, Inc.
doi:10.7551/mitpress/7503.001.0001

For me, that DOI resolves to a 2006 proceedings, and I don't see the referenced article listed.  I would think there's a DOI directly to the article.
