### 2019-06-24 0ebe1af
Fix some DOI's and URL's in JOSS paper
