+++ paper.bib
@@ -32,7 +32,7 @@
   publisher = "Chapman and Hall",
   series = "CRC Interdisiplinary Statistics",
   isbn = {9780412055515},
-  doi = {10.1080/09332480.2014.914768}
+  doi = {10.1201/b14835}
 }
 
 @book{Hastie:2011,
@@ -55,30 +55,14 @@
 }
 
 @inproceedings{Wipf:2007,
-  author    = {David P. Wipf and
-               Srikantan S. Nagarajan},
-  title     = {A New View of Automatic Relevance Determination},
-  booktitle = {NIPS},
-  year      = {2007},
-  ee        = {http://books.nips.cc/papers/files/nips20/NIPS2007_0976.pdf},
-  crossref  = {DBLP:conf/nips/2007},
-  bibsource = {DBLP, http://dblp.uni-trier.de}
-}
-
-@proceedings{DBLP:conf/nips/2007,
-  editor    = {John C. Platt and
-               Daphne Koller and
-               Yoram Singer and
-               Sam T. Roweis},
-  title     = {Advances in Neural Information Processing Systems 20, Proceedings
-               of the Twenty-First Annual Conference on Neural Information
-               Processing Systems, Vancouver, British Columbia, Canada,
-               December 3-6, 2007},
-  booktitle = {NIPS},
-  publisher = {Curran Associates, Inc.},
-  year      = {2008},
-  bibsource = {DBLP, http://dblp.uni-trier.de},
-  doi       = {10.7551/mitpress/7503.001.0001}
+  title = "A New View of Automatic Relevance Determination",
+  author = "David P. Wipf and Srikantan S. Nagarajan",
+  booktitle = "Advances in Neural Information Processing Systems 20",
+  editor = "J. C. Platt and D. Koller and Y. Singer and S. T. Roweis",
+  pages = "1625--1632",
+  year = "2008",
+  publisher = "Curran Associates, Inc.",
+  url = {http://papers.nips.cc/paper/3372-a-new-view-of-automatic-relevance-determination.pdf}
 }
 
 @article{Tibshirani:1996,
