### 2019-06-24 #1359 @jdrugo
@usethedata Thank you for checking the DOI's/URL's and sorry for the mix-up. The first seemed to be a copy/paste error. The second was an error on the NeurIPS webpage. There is no DOI for the article itself (there rarely are for NeurIPS papers), but I have added the URL. They should now be both fixed.
