### 2019-06-25 449d6c3
Add JOSS badge, and link to v0.3
