### 2019-06-25 aea9f95
Add Zenodo DOI badge to README.md
