### 2019-06-25 #1359 @jdrugo
@usethedata I tagged a new version/release v0.3. The Zenodo DOI is https://doi.org/10.5281/zenodo.3255135.
