### 2019-06-26 #1359 @usethedata
@openjournals/joss-eics This looks ready to go to me.  Whedon does suggest 10.1080/09332480.2014.914768 as a DOI for one of the references.  However, when I check that DOI, it resolves to a Taylor and Francis page which is a _review_ of that book, rather than the book itself.  So, I think that the DOI suggestion should not be used.
