### 2019-06-26 #1359 @kyleniemeyer
Hi @jdrugo, paper looks good overall, I have just a few minor edits for you:

- Can you add more complete affiliation information for yourself, similar to what you'd typically see in a journal article? (Department, institution, city, state, country)
- in the third paragraph from the end (beginning with "More specifically, the models...", one of the sentences is missing something: "... where it (hyper-)parameters are tuned...". Perhaps that should be "it's (hyper-)parameters" ?
- in the second to last paragraph, should "Neither of the variants without variable selection use and infer hyper-priors." instead be "Neither of the variants without variable selection use *or* infer hyper-priors" ? (bolded just to indicate the change).
- in the same paragraph, last sentence should be "... and the advantages of either *method* remain to be clarified" (methods -> method)

In the references,

- I think the Drugowitsch 2013 arXiv reference should be changed to:
```
@misc{Drugowitsch:2013,
  author = "Jan Drugowitsch",
  title = "Variational Bayesian inference for linear and logistic regression",
  year = 2013,
  howpublished = {\href{https://arxiv.org/abs/1310.5438}{\tt arXiv:1310.5438 [stat.ML]}},
}
```
- there are also a handful of words in journal/book titles that need `{ }` surrounding them to protect capitalization, like Gibbs, Monte Carlo, Fisher, Romanian, Bayesian (in a few places)
