### 2019-06-27 2ad9db5
Fixed typos and bibliography in JOSS paper
