+++ paper.bib
@@ -1,9 +1,9 @@
 
 @phdthesis{Beal:2003,
   author = "Matthew J. Beal",
-  title = "Variational Algorithms for Approximate Bayesian Inference",
+  title = "Variational Algorithms for Approximate {Bayesian} Inference",
   year = "2003", 
-  school = "Gatsby Computational Neuroscience Unit, University College London",
+  school = {Gatsby Computational Neuroscience Unit, University College London},
   doi = {10.1214/06-ba126}
 }
 
@@ -17,20 +17,20 @@
 
 @book{Gelman:2013,
   author = "Andrew Gelman and John B. Carlin and Hal S. Sern and David B. Dunson and Aki Vehtari and Donald B. Rubin",
-  title = "Bayesian Data Analysis",
+  title = "{Bayesian} Data Analysis",
   edition = "3rd",
   year = "2013",
-  publisher = "Chapman and Hall",
-  series = "CRC Texts in Statistical Science",
+  publisher = {Chapman and Hall},
+  series = {CRC Texts in Statistical Science},
   isbn = {9781439840955}
 }
 
 @book{Gilks:1995,
   editor = "W. R. Gilks and S. Richardson and David Spiegelhalter",
-  title = "Markov Chain Monte Carlo in Practice",
+  title = {Markov Chain Monte Carlo in Practice},
   year = "1995",
-  publisher = "Chapman and Hall",
-  series = "CRC Interdisiplinary Statistics",
+  publisher = {Chapman and Hall},
+  series = {CRC Interdisiplinary Statistics},
   isbn = {9780412055515},
   doi = {10.1201/b14835}
 }
@@ -41,7 +41,7 @@
   edition = "2nd",
   year = "2011",
   publisher = "Springer-Verlag",
-  series = "Springer Series in Statistics",
+  series = {Springer Series in Statistics},
   doi = {10.1007/978-0-387-84858-7}
 }
 
@@ -49,8 +49,8 @@
   author = "Kevin P. Murphy",
   title = "Machine Learning: A Probabilistic Perspective",
   year = "2012",
-  publisher = "The MIT Press",
-  series = "Adaptive Computation and Machine Learning Series",
+  publisher = {The MIT Press},
+  series = {Adaptive Computation and Machine Learning Series},
   isbn = {9780262306164}
 }
 
@@ -61,14 +61,14 @@
   editor = "J. C. Platt and D. Koller and Y. Singer and S. T. Roweis",
   pages = "1625--1632",
   year = "2008",
-  publisher = "Curran Associates, Inc.",
+  publisher = {Curran Associates, Inc.},
   url = {http://papers.nips.cc/paper/3372-a-new-view-of-automatic-relevance-determination.pdf}
 }
 
 @article{Tibshirani:1996,
   author = "Robert Tibshirani",
-  title = "Regression Shrinkage and Selection via the Lasso",
-  journal = "Journal of the Royal Statistical Society B",
+  title = "Regression Shrinkage and Selection via the {Lasso}",
+  journal = {Journal of the Royal Statistical Society B},
   year = "1996",
   pages = "267--288",
   volume = "58",
@@ -78,8 +78,8 @@
 
 @article{MacKay:1992,
   author = "David J. C. MacKay",
-  title = "Bayesian Interpolation",
-  journal = "Neural Computation",
+  title = "{Bayesian} Interpolation",
+  journal = {Neural Computation},
   year = "1992",
   volume = "4",
   number = "3",
@@ -89,8 +89,8 @@
 
 @article{George:1993,
     author = "Edward I. George and Robert E. {McCulloch}",
-    title = "Variable Selection via Gibbs Sampling",
-    journal = "Journal of the American Statistical Association",
[…]
