### 2019-06-27 #1359 @jdrugo
Hi @kyleniemeyer , I have updated `paper.md` and `paper.bib` to fix those issues. Please let me know if there are any issues remaining.
