### 2019-06-27 #1359 @kyleniemeyer
@arfon, in case you didn't get/see the tag in https://github.com/openjournals/joss-papers/pull/799, it looks like there is no citation list in the CrossRef XML. The paper has a good-sized list of references with DOIs, so that seems wrong...
