### 2019-06-27 #1359 @kyleniemeyer
Congrats on your article's publication in JOSS, @jdrugo! Thanks to @ManuelaS for reviewing and  @usethedata for editing (and reviewing)!
